+++ lib/factory_bot/instrumentation/engine.rb
@@ -8,6 +8,11 @@ module FactoryBot
 
       # Fill in some dynamic settings (application related)
       initializer 'factory_bot_instrumentation.config' do
+        # Ensure the FactoryBot gem loads its factories to ensure they are
+        # also available in the rails console and other places in the app
+        # and not only via instrumentation frontend.
+        FactoryBot.reload
+
         FactoryBot::Instrumentation.configure do |conf|
           # Set the application name dynamically
           conf.application_name \